modified:   CMakeLists.txt 	modified:   src/blitz-long/blitz/applics.h 	modified:   src/blitz-long/blitz/array-impl.h 	modified:   src/blitz-long/blitz/array-old.h 	modified:   src/blitz-long/blitz/array.h 	modified:   src/blitz-long/blitz/array/convolve.h 	modified:   src/blitz-long/blitz/array/domain.h 	modified:   src/blitz-long/blitz/array/et.h 	modified:   src/blitz-long/blitz/array/expr.h 	modified:   src/blitz-long/blitz/array/funcs.h 	modified:   src/blitz-long/blitz/array/functorExpr.h 	modified:   src/blitz-long/blitz/array/indirect.h 	modified:   src/blitz-long/blitz/array/iter.h 	modified:   src/blitz-long/blitz/array/newet-macros.h 	modified:   src/blitz-long/blitz/array/newet.h 	modified:   src/blitz-long/blitz/array/ops.h 	modified:   src/blitz-long/blitz/array/reduce.h 	modified:   src/blitz-long/blitz/array/stencilops.h 	modified:   src/blitz-long/blitz/array/stencils.h 	modified:   src/blitz-long/blitz/bench.h 	modified:   src/blitz-long/blitz/benchext.h 	modified:   src/blitz-long/blitz/blitz.h 	modified:   src/blitz-long/blitz/bzconfig.h 	modified:   src/blitz-long/blitz/compiler.h 	modified:   src/blitz-long/blitz/extremum.h 	modified:   src/blitz-long/blitz/funcs.h 	modified:   src/blitz-long/blitz/generate/arroperands.h 	modified:   src/blitz-long/blitz/generate/bzfstream.h 	modified:   src/blitz-long/blitz/generate/operands.h 	modified:   src/blitz-long/blitz/generate/operands2.h 	modified:   src/blitz-long/blitz/indexexpr.h 	modified:   src/blitz-long/blitz/limits-hack.h 	modified:   src/blitz-long/blitz/matexpr.h 	modified:   src/blitz-long/blitz/mathf2.h 	modified:   src/blitz-long/blitz/matrix.h 	modified:   src/blitz-long/blitz/memblock.h 	modified:   src/blitz-long/blitz/meta/dot.h 	modified:   src/blitz-long/blitz/meta/matmat.h 	modified:   src/blitz-long/blitz/meta/matvec.h 	modified:   src/blitz-long/blitz/meta/product.h 	modified:   src/blitz-long/blitz/meta/sum.h 	modified:   src/blitz-long/blitz/minmax.h 	modified:   src/blitz-long/blitz/mstruct.h 	modified:   src/blitz-long/blitz/numinquire.h 	modified:   src/blitz-long/blitz/numtrait.h 	modified:   src/blitz-long/blitz/ops.h 	modified:   src/blitz-long/blitz/promote.h 	modified:   src/blitz-long/blitz/rand-dunif.h 	modified:   src/blitz-long/blitz/rand-mt.h 	modified:   src/blitz-long/blitz/rand-normal.h 	modified:   src/blitz-long/blitz/rand-tt800.h 	modified:   src/blitz-long/blitz/rand-uniform.h 	modified:   src/blitz-long/blitz/random.h 	modified:   src/blitz-long/blitz/range.h 	modified:   src/blitz-long/blitz/reduce.h 	modified:   src/blitz-long/blitz/timer.h 	modified:   src/blitz-long/blitz/tiny.h 	modified:   src/blitz-long/blitz/tinymat.h 	modified:   src/blitz-long/blitz/tinyvec-et.h 	modified:   src/blitz-long/blitz/tinyvec.h 	modified:   src/blitz-long/blitz/tinyveciter.h 	modified:   src/blitz-long/blitz/traversal.h 	modified:   src/blitz-long/blitz/tvecglobs.h 	modified:   src/blitz-long/blitz/update.h 	modified:   src/blitz-long/blitz/vecexpr.h 	modified:   src/blitz-long/blitz/vecexprwrap.h 	modified:   src/blitz-long/blitz/vecglobs.h 	modified:   src/blitz-long/blitz/vecpick.h 	modified:   src/blitz-long/blitz/vecpickiter.h 	modified:   src/blitz-long/blitz/vector-et.h 	modified:   src/blitz-long/blitz/vector.h 	modified:   src/blitz-long/blitz/vecwhere.h 	modified:   src/blitz-long/blitz/zero.h 	modified:   src/common/CMakeLists.txt 	modified:   src/common/common.cpp 	modified:   src/common/common.h 	modified:   src/common/experiment.cpp 	modified:   src/common/experiment.h 	modified:   src/common/flatfield.h 	modified:   src/common/ipc.h 	modified:   src/common/kernel.cpp 	modified:   src/common/lipc.cpp 	modified:   src/executables/CMakeLists.txt 	modified:   src/executables/ct-abs.cpp 	modified:   src/executables/f2i.cpp 	modified:   src/executables/ff.cpp 	modified:   src/executables/ipc.cpp 	modified:   src/executables/sino-abs.cpp 	modified:   src/executables/sino.cpp 	modified:   src/executables/stitch.cpp

+++ src/common/kernel.cpp
@@ -563,7 +563,7 @@ CTrec::reconstruct_uni(Map &sinogram, Map &result, float center) const {
 
   const int zShift = (int)(_pixels*(zPad-1)/2);
   const int thetas = sinogram.rows();
-  Line zsinoline((int)(_pixels*zPad)); // zero-padded sinoline.
+  Line zsinoline(_pixels*zPad); // zero-padded sinoline.
 
   if ( _contrast != Contrast::FLT ) {
     for (int iTheta = 0 ; iTheta < thetas ; iTheta++) {
@@ -582,6 +582,7 @@ CTrec::reconstruct_uni(Map &sinogram, Map &result, float center) const {
   }
 
   project_sino(sinogram, result, center, _threads);
+
 
 }
 
@@ -974,7 +975,7 @@ ts_add( Map &projection, Map &result, const Filter & filter,
   float plane_cos = (plane-center)*cur_sin;
   float Rcenter = pixels*0.5 + center;
 
-  for (int ycur = 0 ; ycur < thetas ; ycur++){
+  for (long ycur = 0 ; ycur < thetas ; ycur++) {
 
 	Line ln = projection(ycur, blitz::Range::all());
 
@@ -984,12 +985,10 @@ ts_add( Map &projection, Map &result, const Filter & filter,
 	  partial_sum( ln.begin(), ln.end(), ln.begin() );
 
 	// projecting
-	for (int xcur = 0 ; xcur < pixels ; xcur++) {
+	for (long xcur = 0 ; xcur < pixels ; xcur++) {
 	  float di = ( xcur - Rcenter )*cur_cos - plane_cos + Rcenter;
-	  di = (di < 0) ?
-		0 :
-		(di >=pixels ) ? pixels - 1 : di;
-	  result(ycur, xcur) += projection(ycur, int(di));
+	  di = (di < 0)  ?  0  :  (di >=pixels ) ? pixels - 1 : di;
+	  result(ycur, xcur) += projection(ycur, long(di));
 	}
 
   }
